+++ APISunSale/Controllers/QuestoesController.cs
@@ -173,6 +173,15 @@ namespace APISunSale.Controllers
             {
                 var user = await _utils.GetUserFromContextAsync();
 
+                if (user.Admin != "1")
+                {
+                    return new ResponseBase<MainViewModel>()
+                    {
+                        Message = "Acesso não autorizado",
+                        Success = false
+                    };
+                }
+
                 var result = await _service.Add(_mapper.Map<MainEntity>(main), user.Id);
                 return new ResponseBase<MainViewModel>()
                 {
@@ -200,7 +209,18 @@ namespace APISunSale.Controllers
         {
             try
             {
-                if(await _service.GetById(main.Codigo) == null)
+                var user = await _utils.GetUserFromContextAsync();
+
+                if (user.Admin != "1")
+                {
+                    return new ResponseBase<MainViewModel>()
+                    {
+                        Message = "Acesso não autorizado",
+                        Success = false
+                    };
+                }
+
+                if (await _service.GetById(main.Codigo) == null)
                 {
                     return new ResponseBase<MainViewModel>()
                     {
@@ -237,6 +257,17 @@ namespace APISunSale.Controllers
         {
             try
             {
+                var user = await _utils.GetUserFromContextAsync();
+
+                if (user.Admin != "1")
+                {
+                    return new ResponseBase<bool>()
+                    {
+                        Message = "Acesso não autorizado",
+                        Success = false
+                    };
+                }
+
                 var result = await _service.DeleteById(id);
                 return new ResponseBase<bool>()
                 {
@@ -384,6 +415,16 @@ namespace APISunSale.Controllers
             try
             {
                 var user = await _utils.GetUserFromContextAsync();
+
+                if(user.Admin != "1")
+                {
+                    return new ResponseBase<List<MainViewModel>>()
+                    {
+                        Message = "Acesso não autorizado",
+                        Success = false
+                    };
+                }
+
                 List<MainViewModel> list = new List<MainViewModel>();
                 List<string> linhas = new List<string>();
                 string materia = "DIREITO PREVIDENCIÁRIO";
